requirements.txt
SpeechRecognition
google-generativeai
google-colab
pyautogui